code/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.16.2
openai==1.3.5
python-multipart==0.0.6
python-docx==1.1.0
pypdf2==3.0.1
requests==2.31.0